legacy/requirements_backup.txt
requests
beautifulsoup4
pandas
matplotlib
supabase
python-dotenv
